LPS-165445 Merge IndexerRequestBufferExecutor and BaseIndexerRequestBufferExecutor into DefaultIndexerRequestBufferExecutor and remove property setting as it is the only implementation
